src/test/java/com/nocommerce/demo/testsuite/LoginPageTest.java
package com.nocommerce.demo.testsuite;

import com.nocommerce.demo.customelisteners.CustomListeners;
import com.nocommerce.demo.pages.HomePage;
import com.nocommerce.demo.pages.LoginPage;
import com.nocommerce.demo.testbase.BaseTest;
import org.testng.Assert;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Listeners;
import org.testng.annotations.Test;

@Listeners(CustomListeners.class)
public class LoginPageTest extends BaseTest {

    HomePage homePage;
    LoginPage loginPage;
    RegisterPageTest registerPageTest;

    @BeforeMethod (alwaysRun = true)
    public void inIt(){
        homePage = new HomePage();
        loginPage = new LoginPage();
        registerPageTest = new RegisterPageTest();
    }

    @Test(groups = {"sanity", "smoke", "regression"})
    public void userShouldNavigateToLoginPageSuccessFully(){
        homePage.clickOnLoginLink();
        String expectedMessage = "Welcome, Please Sign In!";
        String actualMessage = loginPage.getWelcomeText();
        Assert.assertEquals(actualMessage, expectedMessage, "Login page not displayed");
    }

    @Test(groups = {"smoke", "regression"})
    public void verifyTheErrorMessageWithInValidCredentials(){
        homePage.clickOnLoginLink();
        loginPage.loginToTheApplication("abc@gmail.com", "abc123");
        String expectedMessage = "Login was unsuccessful. Please correct the errors and try again.\n" +
                "No customer account found";
        String actualMessage = loginPage.getLoginUnsuccessfulMessage();
        Assert.assertEquals(actualMessage, expectedMessage, "Login Unsuccessful Message Not Displayed");
    }

    @Test(groups = {"regression"})
    public void verifyThatUserShouldLogInSuccessFullyWithValidCredentials(){
        homePage.clickOnLoginLink();
        loginPage.loginToTheApplication("abcdxyz@gmail.com", "abcd123");
        Assert.assertTrue(homePage.isLogoutLinkAvailable(),"LogOut link is not available");
    }

    @Test(groups = {"regression"})
    public void verifyThatUserShouldLogOutSuccessFully() {
        homePage.clickOnLoginLink();
        loginPage.loginToTheApplication("abcdxyz@gmail.com", "abcd123");
        homePage.clickOnLogout();
        homePage.explicitWait(15);
        Assert.assertTrue(homePage.isLoginLinkAvailable(), "Login link is not available");
    }
}
